Master is ACTUALLY functional now
Fixed the many dependency errors we were having across platforms, and
finally figure out the reason (websocket-client is 100% different than
websocket, even if they have the same module name...). Added a fix.sh
file for linux users that had this issue. It's simply purging and
reinstalling the correct versions. Also updated the ReadMe to properly
reflect setup, as well as the Help.py which has been outdated for a bit.